[HUDI-994] Move TestHoodieIndex test cases to unit tests
